refactoring watchdog
